tests/requirements.txt
pytest==3.0.3
pyyaml==3.12
